implements auth with jwt

+++ js-problem-solver-backend/app/controllers/problems_controller.rb
@@ -1,4 +1,7 @@
 class ProblemsController < ApplicationController
+    include ActionController::HttpAuthentication::Token
+    before_action :authenticate_user, only: [:create, :destroy]
+
     def index
         render json: ProblemSerializer.new(Problem.all)
     end
@@ -7,6 +10,8 @@ class ProblemsController < ApplicationController
         problem = Problem.create(problem_params)
         if problem.save
             render json: ProblemSerializer.new(problem)
+        else
+            render json: {error: "Problem did not save."}
         end
     end
 
@@ -19,13 +24,22 @@ class ProblemsController < ApplicationController
 
     def destroy
         if problem = Problem.find(params[:id])
-            problem.destroy
+            # binding.pry
+            problem.destroy if session[:user_id] === problem.user_id
         else
             render json: {message: "Error! Try again."}
         end
     end
 
     private
+
+    def authenticate_user
+        token, _options = token_and_options(request)
+        user_id = SessionsController.decode(token)
+        User.find(user_id)
+    rescue ActiveRecord::RecordNotFound
+        render json: {authorized: false}
+    end
 
     def problem_params
         params.require(:problem).permit(:title, :description, :user_id)
